requirement.txt
# pip install - r requirement.txt
# streamlit run frontend.py

streamlit
torch
torchvision
pillow
numpy
matplotlib
scikit-learn
tqdm
torchsummary